requirements.txt
torch
torchvision
streamlit
Pillow
gdown
pandas
plotly
reportlab